Address review: merge the three electron-energy test scripts into one (--case), default qdsmc_n_floor to n_floor, trim a change-relative comment
