+++ Docs/source/usage/parameters.rst
@@ -3740,11 +3740,11 @@ Maxwell solver: kinetic-fluid hybrid
 
 .. pp:param:: hybrid_pic_model.qdsmc_n_floor
     :type: ``float``
-    :default: ``1``
+    :default: :pp:param:`hybrid_pic_model.n_floor`
     :optional:
 
     Density floor, in :math:`m^{-3}`, below which cells are excluded from the QDSMC electron-energy-equation
-    update (the electron temperature is left unchanged there).
+    update (the electron temperature is left unchanged there). Defaults to :pp:param:`hybrid_pic_model.n_floor`.
 
 .. pp:param:: hybrid_pic_model.include_joule_heating
     :type: ``bool``
