+++ Examples/Tests/ohm_solver_electron_energy_eq/CMakeLists.txt
@@ -5,7 +5,7 @@ add_warpx_test(
     test_2d_ohm_solver_electron_energy_adiabat_picmi  # name
     2  # dims
     2  # nprocs
-    "inputs_test_2d_ohm_solver_electron_energy_adiabat_picmi.py --test"  # inputs
+    "inputs_test_2d_ohm_solver_electron_energy_picmi.py --case adiabat --test"  # inputs
     "analysis_adiabat.py --tol-median 0.01 --tol-max 0.05"  # analysis
     "analysis_default_regression.py --path diags/field_diags"  # checksum
     OFF  # dependency
@@ -15,7 +15,7 @@ add_warpx_test(
     test_2d_ohm_solver_electron_energy_joule_picmi  # name
     2  # dims
     2  # nprocs
-    "inputs_test_2d_ohm_solver_electron_energy_joule_picmi.py --test --eta-scale 100"  # inputs
+    "inputs_test_2d_ohm_solver_electron_energy_picmi.py --case joule --test --eta-scale 100"  # inputs
     "analysis_joule.py --eta-scale 100"  # analysis
     "analysis_default_regression.py --path diags/field_diags"  # checksum
     OFF  # dependency
@@ -25,7 +25,7 @@ add_warpx_test(
     test_2d_ohm_solver_electron_energy_qei_picmi  # name
     2  # dims
     2  # nprocs
-    "inputs_test_2d_ohm_solver_electron_energy_qei_picmi.py --test"  # inputs
+    "inputs_test_2d_ohm_solver_electron_energy_picmi.py --case qei --test"  # inputs
     "analysis_qei.py --nu-ei 1e6"  # analysis
     "analysis_default_regression.py --path diags/field_diags"  # checksum
     OFF  # dependency
